+++ lib/hirefire/client.rb
@@ -44,8 +44,6 @@ module HireFire
       when Net::HTTPServerError
         raise RequestError, "Server responded with #{response.code} status."
       else
-        return :payload_too_large if response.code.to_i == 413
-
         raise RequestError, "Unexpected response code #{response.code}."
       end
     end